client is working!

+++ request.go
@@ -2,11 +2,10 @@ package yar
 
 type Request struct {
 	Protocol *Protocol   `json:"-" msgpack:"-"`
-	Id       int64       `json:"i" msgpack:"i"`
+	Id       uint32       `json:"i" msgpack:"i"`
 	Method   string      `json:"m" msgpack:"m"`
 	Params   interface{} `json:"p" msgpack:"p"`
 }
-
 
 func (self *Request)getParam(index int,v interface{}) {
